Create AUTOMATIC LIQUID DISPENSER
This it to all my team mates. Since most of the group are working on the automatic sanitizing and alcohol dispensing machines I feel like we should be doing something unique. So my suggestion her is to make an automatic juice dispenser machine. I mean think about it , we get to make something unique and we also get to drink juice UwU. I'll share the link to my project plan in this file. check it out if you want to.